// src/components/merch-widget/types.ts
export type Color = {
  id: string;
  name: string;
  hex: string;
};

export type PlacementZone = {
  id: string;
  name: string;
  top: string;
  left: string;
  width: string;
};

export type Product = {
  id: string;
  name: string;
  price: number;
  colors: Color[];
  zones: PlacementZone[];
  image: string;
};

export type LogoState = {
  url: string | null;
  scale: number;
  flipH: boolean;
  flipV: boolean;
};